backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pydantic>=2.8.0
python-dotenv==1.0.1
sqlalchemy==2.0.36
aiosqlite==0.20.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==3.2.2
pillow>=10.4.0
opencv-python-headless>=4.9.0.80
numpy>=2.1.0
scipy>=1.14.0
scikit-learn>=1.6.0
matplotlib>=3.9.0
albumentations>=1.4.8
timm>=1.0.7